part 4.23 done

+++ part4/controllers/blogs.js
@@ -1,32 +1,72 @@
 const blogsRouter = require('express').Router()
+//const _ = require('lodash')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
 
 
 blogsRouter.get('/', async (request, response) => {
-  const notes = await Blog.find({})
-  response.json(notes.map(blog => blog.toJSON()))
+  const blogs = await Blog.find({})
+  response.json(blogs.map(blog => blog.toJSON()))
+})
+
+
+blogsRouter.get('/:id', async (req, res) => {
+  const blog = await Blog.findById(req.params.id).populate('user', {
+    username: 1,
+    name: 1,
+  })
+  if (blog) {
+    res.json(blog.toJSON())
+  } else {
+    return res.status(404).end()
+  }
 })
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  if (body.title === undefined || body.url === undefined) {
-    response.status(400).end()
-  } else{
-    const blog = new Blog({
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: body.likes || 0,
-    })
-    const savedBlog = await blog.save()
-    response.json(savedBlog)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
   }
+  const user = await User.findById(decodedToken.id)
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes || 0,
+    user: user._id
+  })
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.json(savedBlog.toJSON())
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else {
+    return response.status(401).json({ error: 'tokens dont match' })
+  }
 })
+
 
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
